# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plumbing Business AI Platform API
  description: |
    Comprehensive API for the production-ready plumbing business AI platform.
    
    This API provides endpoints for managing customers, jobs, conversations, AI interactions,
    emergency responses, scheduling, pricing, warranties, and system monitoring.
    
    ## Features
    - **Customer Management**: Complete customer lifecycle management
    - **AI-Powered Conversations**: Intelligent message processing and response generation
    - **Emergency Response**: Real-time emergency detection and routing
    - **Dynamic Pricing**: Intelligent pricing based on multiple factors
    - **Scheduling Optimization**: Advanced technician scheduling and route optimization
    - **Warranty Management**: Comprehensive warranty tracking and claims processing
    - **Performance Monitoring**: Real-time system health and performance metrics
    
    ## Authentication
    All endpoints require JWT-based authentication unless otherwise specified.
    Include the token in the Authorization header: `Bearer {token}`
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - Standard endpoints: 1000 requests per hour per user
    - Search endpoints: 100 requests per minute per user
    - Webhook endpoints: 10,000 requests per hour
    
    ## Error Handling
    The API uses standard HTTP status codes and returns structured error responses:
    ```json
    {
      "error": {
...
        }
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /webhooks/google-voice:
    post:
      tags: [Webhooks]
      summary: Google Voice webhook
      description: Handle incoming messages and calls from Google Voice
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, eventType]
              properties:
                message:
                  type: object
                  properties:
                    text:
                      type: string
                    phoneNumber:
                      type: string
                    timestamp:
                      type: string
                      format: date-time
                    type:
                      type: string
                      enum: [sms, voice_call, voicemail]
                    direction:
                      type: string
                      enum: [inbound, outbound]
                eventType:
                  type: string
                  enum: [message_received, call_received, voicemail_received]
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  processed:
                    type: boolean
                  conversationId:
                    type: string
                    format: uuid
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...